Split lua functions into python dictionaries for each firmware.
Corrected a bug that wouldn't allow quality or materials profile creation.

Finished the overall printer.lua creator for Marlin
