+++ style.tcss
@@ -1,6 +1,6 @@
 .app-grid {
     layout: grid;
-    grid-size: 3;  /* three columns */
+    grid-size: 3 2;  /* three columns */
     grid-columns: 1fr;
     grid-rows: 1fr;
     grid-gutter: 1;
@@ -14,10 +14,11 @@
 
 .app-grid-2 {
     layout: grid;
-    grid-size: 4;  /* 4 columns 2 rows */
+    grid-size: 4 13;  /* 4 columns 2 rows */
     grid-columns: 1fr;
     grid-rows: 1fr;
     grid-gutter: 1;
+    background: $panel;
 }
 
 .features > Static {
@@ -31,17 +32,50 @@
     width: 100%;
     height: 100%;
     column-span: 2;
+    row-span: 12;
     background: $panel;
     border: #272635;
 }
 
+.features-2 > Static {
+    overflow-y: auto;
+    background: $boost;
+    color: auto;
+    margin-bottom: 1;
+    padding: 1;
+}
+.features-2 {
+    width: 100%;
+    height: 100%;
+    column-span: 1;
+    row-span: 12;
+    background: $panel;
+    border: #272635;
+}
+
+#static_advanced_mode {
+    color: rgb(200, 0, 0);
+}
+
+#static_toggle_code_modification {
+    color: rgb(200, 0, 0);
+}
+
+.important_toggle{
+    height: auto;
+    width: 32%;
+}
+
+.important_toggle> .switch--slider {
+    color: rgb(200, 0, 0);
+}
 
 .event-text {
     overflow-x: auto;
     overflow-y: auto;
     height: 100%;
     column-span: 2;
-    row-span: 2;
+    row-span: 12;
     background: $panel;
     border: rgb(0, 40, 100);
     
@@ -61,8 +95,8 @@
     overflow-x: auto;
     overflow-y: auto;
     height: 100%;
-    column-span: 2;
-    row-span: 2;
+    column-span: 3;
+    row-span: 12;
     background: $panel;
     border: rgb(0, 40, 100);
     
@@ -71,13 +105,23 @@
 .event-text-2 > Static {
     overflow-x: auto;
     overflow-y: auto;
-    width: auto;
-    height: auto;
+    height: 3;
     margin-right: 1;
-    background: $boost;
-    
+    content-align: center middle;
+    width: 100%;
+    background: rgb(0, 40, 100);
+
 }
 
+.event-text-2 > TextArea {
+    overflow-x: auto;
+    overflow-y: auto;
+    height: auto;
+    content-align: center middle;
+    width: 100%;
+    background: $panel;
+
+}
 .container {
     layout: horizontal;
     height: 3;
@@ -99,25 +143,27 @@ Button {
     width: 100%;
 }
 
-.label {
-    height: 3;
-    content-align: center middle;
-    width: 100%;
-}
-
 .feature-text {
     height: 3;
     background: rgb(0, 0, 0);
     content-align: left middle;
     width: 68%;
 }
+
 Input {
     width: 32%;
     background: rgb(0, 40, 100);
 }
 
-
 .horizontal-layout {
     layout: horizontal;
     height: auto;
+}
+
+.warning {
+    column-span:4;
+    row-span: 1;
+    height: 2;
+    width: 100%;
+    background: rgb(0, 40, 100);
 }
