+++ testProfile/materials/PLA.lua
@@ -1,0 +1,26 @@
+--Custom material profile: PLA for testProfile 
+--Created on 06/05/24
+ 
+
+ 
+--extruder_changes
+nozzle_diameter_mm_0 = 0.4
+filament_diameter_mm_0 = 1.75
+filament_linear_adv_factor = 0.06
+ 
+--retraction_changes
+filament_priming_mm = 45
+ 
+--temperature_changes
+extruder_temp_degree_c = 210
+bed_temp_degree_c = 60
+--chamber_temp_degree_c = 0
+enable_fan = true
+fan_speed_percent = 100
+fan_speed_percent_on_bridges = 100
+ 
+--speed_changes
+print_speed_mm_per_sec = 60
+perimeter_print_speed_mm_per_sec = 45
+cover_print_speed_mm_per_sec = 30
+first_layer_print_speed_mm_per_sec = 20
