+++ testProfile/printer.lua
@@ -1,10 +1,11 @@
 --Printer functions for testProfile
---Created on 06/04/24
+--Created on 06/05/24
  
 output(';FLAVOR:Marlin') 
-output(';Layer height: ' .. round(z_layer_height_mm,2)) 
-output(';Generated with ' .. slicer_name .. ' ' .. slicer_version .. '\n')--//////////////////////////////////////////////////Defining main variables 
-
+output(';Layer height: ' .. round(z_layer_height_mm, 2)) 
+output(';Generated with ' .. slicer_name .. ' ' .. slicer_version .. '\n') 
+ 
+--//////////////////////////////////////////////////Defining main variables
 extruder_e = 0
 exruder_e_restart = 0
 current_z = 0.0
@@ -41,11 +42,15 @@ function header()
 
     output('M109 R' .. extruder_temp_degree_c[extruders[0]] .. ' ; set extruder temp')
     output('M190 S' .. bed_temp_degree_c .. ' ; wait for bed temp')
-    
+        
     output('M107')
     output('G28 ; home all without mesh bed level')
 
-    
+        
+    --start auto bed leveling and reload previous bed mesh
+    output('M420 S1 ; enable bed leveling (was disabled y G28)')
+    output('M420 L ; load previous bed mesh')
+        
     output('M109 S' .. extruder_temp_degree_c[extruders[0]] .. ' ; wait for extruder temp')
 
     output('')
@@ -56,7 +61,7 @@ function header()
     changed_frate = true
 
 end
------------------------
+        -----------------------
 function footer()
     --called to create the footer of the G-Code file.
 
@@ -64,7 +69,7 @@ function footer()
     output('G4 ; wait')
     output('M104 S0 ; turn off temperature')
     output('M140 S0 ; turn off heatbed')
-    
+  
 
     output('M107 ; turn off fan')
     output('G28 X Y ; home X and Y axis')
@@ -73,15 +78,9 @@ function footer()
     output('G90')
     output('M84 ; disable motors')
     output('')
-    
-    --set limits back to original values.
-    output('M201 X' .. x_max_acc .. ' Y' .. y_max_acc .. ' Z' .. z_max_acc .. ' E' .. e_max_acc .. ' ; sets maximum accelerations, mm/sec^2')
-    output('M203 X' .. x_max_speed .. ' Y' .. y_max_speed .. ' Z' .. z_max_speed .. ' E' .. e_max_speed .. ' ; sets maximum feedrates, mm/sec')
-    output('M204 P' .. default_acc .. ' R' .. e_prime_max_acc .. ' T' .. default_acc .. ' ; sets acceleration (P, T) and retract acceleration (R), mm/sec^2')
-    output('M205 S0 T0 ; sets the minimum extruding and travel feed rate, mm/sec')
-    output('M205 J' .. default_junction_deviation .. ' ; sets Junction Deviation')
-            
+  
 end
+  
 
 --################################################## COMMENT 
 function comment(text)
@@ -91,7 +90,8 @@ function comment(text)
     output('; ' .. text)
 end
 
--------------------------################################################## LAYER
+-----------------------
+--################################################## LAYER
 function layer_start(zheight)
   --called at the start of a layer at height "zheight" (mm).
                                    
@@ -120,7 +120,8 @@ function layer_stop()
                                   
 end
                                   
--------------------------################################################## EXTRUDER
+-----------------------
+--################################################## EXTRUDER
 function extruder_start()
   --called before extruding.
 
@@ -131,17 +132,8 @@ function extruder_stop()
 
 end
 -----------------------
-function select_extruder(extruder)
---[[
-    called when setting-up the extruder "extruder". This function is called for each available extruder at 
-    the beginning of the G-Code and once for the first used extruder in the print. 
-    After this, IceSL calls "swap_extruder".
-    ]]
-    -- hack to work around not beeing a lua global
     local n = nozzle_diameter_mm_0
-
-    local n = nozzle_diameter_mm_0 -- should be changed to nozzle_diameter_mm[extruder] when available
-    
+  
     local x_pos = 0.1
     local y_pos = 20
     local z_pos = 0.3
@@ -175,14 +167,14 @@ function select_extruder(extruder)
     current_frate = travel_speed_mm_per_sec * 60
     changed_frate = true
   end
-  
   -----------------------
 function swap_extruder(ext1,ext2,x,y,z)
   --[[
     called when swapping extruder 'ext1' to 'ext2' at position x,y,z.
     ]]
 end
--------------------------################################################## MOVEMENTS
+-----------------------
+--################################################## MOVEMENTS
 function prime(extruder,e)
   --[[
     called when priming from value "e" with extruder "extruder". 
@@ -229,16 +221,14 @@ function move_e(e)
     end
 end
 -----------------------
+function move_xyz(x,y,z)      
   --[[
     called when traveling to "x,y,z".
     ]]
     if processing == true then
         processing = false
         output(';travel')
-        if use_per_path_accel then
-            output('M204 S' .. default_acc)
-        end
-    end
+  
   
     if z == current_z then
         if changed_frate == true then
@@ -257,7 +247,8 @@ end
         current_z = z
     end
 end
------------------------
+  -----------------------
+function move_xyze(x,y,z,e)
   --[[
     called when traveling to "x,y,z" while extruding to value "e".
     ]]
@@ -266,30 +257,19 @@ end
     local e_value = extruder_e - extruder_e_restart
   
     if processing == false then
-      processing = true
-      local p_type = 1 -- default paths naming
-      if craftware_debug then p_type = 2 end
-      if      path_is_perimeter then output(path_type[1][p_type])
-      elseif  path_is_shell     then output(path_type[2][p_type])
-      elseif  path_is_infill    then output(path_type[3][p_type])
-      elseif  path_is_raft      then output(path_type[4][p_type])
-      elseif  path_is_brim      then output(path_type[5][p_type])
-      elseif  path_is_shield    then output(path_type[6][p_type])
-      elseif  path_is_support   then output(path_type[7][p_type])
-      elseif  path_is_tower     then output(path_type[8][p_type])
-    end
+        processing = true
   
-      -- acceleration management
-    if use_per_path_accel then
-      if     path_is_perimeter or path_is_shell 
-            then set_acceleration(perimeter_acc, default_jerk)
-      elseif path_is_infill                     
-            then set_acceleration(infill_acc, infill_jerk)
-      elseif (path_is_raft or path_is_brim or path_is_shield or path_is_support or path_is_tower)
-            then set_acceleration(default_acc, default_jerk)
-      end
+        p_type = 2
+  
+        if      path_is_perimeter then output(path_type[1][p_type])
+        elseif  path_is_shell     then output(path_type[2][p_type])
+        elseif  path_is_infill    then output(path_type[3][p_type])
+        elseif  path_is_raft      then output(path_type[4][p_type])
+        elseif  path_is_brim      then output(path_type[5][p_type])
+        elseif  path_is_shield    then output(path_type[6][p_type])
+        elseif  path_is_support   then output(path_type[7][p_type])
+        elseif  path_is_tower     then output(path_type[8][p_type])
     end
-end
   
     if z == current_z then
       if changed_frate == true then
@@ -308,12 +288,14 @@ end
       current_z = z
     end
 end
--------------------------################################################## PROGRESS 
+
+-----------------------
+--################################################## PROGRESS 
 function progress(percent)
   output('M73 P' .. percent)
 end
-
--------------------------################################################## SET
+-----------------------
+--################################################## SET
 function set_feedrate(rate)
   --[[
     called when setting the feed-rate of the printer to "rate".
@@ -340,7 +322,8 @@ function set_extruder_temperature(extruder,temperature)
     ]]
     output('M104 S' .. temperature)
 end
--------------------------################################################## WAIT
+-----------------------
+--################################################## WAIT
 function wait(sec,x,y,z)
   --[[
     called when the parameter "enable_min_layer_time" is set to true
@@ -353,7 +336,8 @@ function wait(sec,x,y,z)
     output("G4 S" .. sec .. "; wait for " .. sec .. "s")
     output("G0 F" .. travel_speed_mm_per_sec .. " X" .. f(x) .. " Y" .. f(y) .. " Z" .. ff(z))
   end
--------------------------################################################## MIXING PARAMETERS
+-----------------------
+--################################################## MIXING PARAMETERS
 function set_and_wait_extruder_temperature(extruder,temperature)
   --[[
     called when setting the extruder "extruder" temperature to "temperature" while waiting. 
@@ -370,3 +354,65 @@ function set_mixing_ratios(ratios)
     ]]
 end        
 -----------------------
+ 
+
+--//////////////////////////////////////////////////Util functions - mainly for unit conversions
+function round(number, decimals)
+  --[[
+    returns a rounded value of "number"
+    ]]
+  local power = 10^decimals
+  return math.floor(number * power) / power
+end
+-----------------------
+function vol_to_mass(volume, density)
+  --[[
+    converts current volume to mass.
+    Used along with the next function to approximate filament consumption."
+    ]]
+  return density * volume
+end
+-----------------------
+function e_to_mm_cube(filament_diameter, e)
+  --[[
+    Uses the filament's dimensions and extrusion width to approximate the volume (mm^3)
+    Used along with the previous function to approximate filament consumption.
+    ]]
+  local r = filament_diameter / 2
+  return (math.pi * r^2 ) * e
+end
+-----------------------
+-- get the E value (for G1 move) from a specified deposition move
+function e_from_dep(dep_length, dep_width, dep_height, extruder)
+  --[[
+    Yet to be understood
+  ]]
+  local r1 = dep_width / 2
+  local r2 = filament_diameter_mm[extruder] / 2
+  local extruded_vol = dep_length * math.pi * r1 * dep_height
+  return extruded_vol / (math.pi * r2^2)
+end
+-----------------------
+function jerk_to_junction_deviation(jerk, accel)
+  --[[
+    Converts Marlin jerk value to junction deviation
+    ]]
+  return 0.4 * ( (jerk^2) / accel )
+end
+-----------------------
+-- marlin jerk * sqrt(2) = square corner velocity
+-- scv = square corner velocity
+function scv_to_jerk(scv) 
+  --[[
+    converts klipper scv value to marlin jerk value
+    ]]
+  return math.sqrt(2) * scv
+end
+-----------------------
+function jerk_to_scv(jerk)
+  --[[
+    converts marlin jerk value to klipper scv value
+    ]]
+  return jerk/math.sqrt(2)
+end
+
