+++ testProfile/profiles/low.lua
@@ -1,0 +1,17 @@
+--Custom quality profile: low for testProfile 
+--Created on 06/05/24
+ 
+
+ 
+--layer_thickness_changes
+z_layer_height_mm = 0.2
+ 
+--speed_changes
+print_speed_mm_per_sec = 60
+perimeter_print_speed_mm_per_sec = 30
+cover_print_speed_mm_per_sec = 35
+first_layer_print_speed_mm_per_sec = 20
+travel_speed_mm_per_sec = 80
+priming_mm_per_sec = 30
+retract_mm_per_sec = 40
+speed_multiplier_0 = 1
