+++ tooltips.py
@@ -1,0 +1,69 @@
+tooltips = {
+    "static_firmware": "Defines G-code flavor. Entirely defines your printer's behaviour and the way your 3d model is translated to G-code instructions understandable by the machine. ",
+    "static_advanced_mode": "Toggle to hide/display advanced features. Please note they are mostly assigned according to parent-features, thus modifying them may yield unexpected results.",
+    "static_build_area_dimensions": "Settings defining your printer's bed",
+    "static_bed_circular": "Defines your bed's shape; it can either be rectangular or circular. \n Having a circular bed means the size parameters are solely defined by the bed's radius.",
+    "static_bed_radius": "If circular, defines your bed's radius and thus, its size-parameters.",
+    "static_bed_size_x_mm": "Your bed's length (mm)",
+    "static_bed_size_y_mm": "Your bed's width (mm)",
+    "static_bed_size_z_mm": "Your bed's height (mm)",
+    "static_extruder_settings": "Settings defining your extruder(s) such as the nozzle diameter.",
+    "static_extruder_count": "The number of extruders on your printer.",
+    "static_nozzle_diameter_mm_0": "Defines your first extruder's nozzle diameter (mm).",
+    "static_filament_diameter_mm_0": "Defines your first extruder's filament diameter (mm).",
+    "static_filament_linear_adv_factor": "Linear Advance grants the firmware the ability to predict the pressure build-up in the extruder at higher speed, therefore allowing it to decrease the flow of material to avoid blobs and other artifacts. This parameter defines the k-factor for said decrease.",
+    "static_retraction_settings": "Optional and advanced settings. Defines your printer's behaviour when it retracts (z-axis) itself from a print.",
+    "static_filament_priming_mm": "Defines the length of the filament retracted to prevent oozing during travel motions.",
+    "static_priming_mm_per_sec": "Defines the speed at which the filament is retracted.",
+    "static_retract_mm_per_sec": "Defines the speed at which the extruder retracts itself (z-axis) from the current print.",
+    "static_layer_height": "Settings defining the thickness of each layer.",
+    "static_z_layer_height_mm": "Defines the general thickness of each layer (mm). It is usually used to qualify the print's overall quality with 0.2 being a draft.",
+    "static_z_layer_height_mm_min": "Defines the smallest thickness (mm) achievable by the printer. Automatically scaled in accordance to nozzle diameter.",
+    "static_z_layer_height_mm_max": "Defines the largest thickness (mm) achievable by the printer. Automatically scaled in accordance to nozzle diameter.",
+    "static_printing_temperatures": "Settings defining your printer's default temperatures (C). Highly dependent on the material.",
+    "static_extruder_temp_degree_c": "Defines your extruder's temperature. Please refer to your filament's recommendations.",
+    "static_extruder_temp_degree_c_min": "Based on the previous setting",
+    "static_extruder_temp_degree_c_max": "Based on the previous setting",
+    "static_bed_temp_degree_c": "Defines your bed's heating temperature. Heating your bed to the correct temperature is crucial to efficiently deposit the first layer and make sure it sticks to the surface.",
+    "static_bed_temp_degree_c_min": "Based on the previous setting.",
+    "static_bed_temp_degree_c_max": "Based on the previous setting.",
+    "static_printing_speeds": "Settings defining the general speed at which your printer moves its extruder around.",
+    "static_print_speed_mm_per_sec": "Defines the general printing speed, meaning the speed at which your nozzle actively extrudes material.",
+    "static_print_speed_mm_per_sec_min": "Based on the previous setting.",
+    "static_print_speed_mm_per_sec_max": "Based on the previous setting. ",
+    "static_perimeter_print_speed_mm_per_sec": "Defines the specific speed at which perimeters are printed. Set by default to 3/4 of the general printing speed.",
+    "static_perimeter_print_speed_mm_per_sec_min": "Based on the previous setting.",
+    "static_perimeter_print_speed_mm_per_sec_max": "Based on the previous setting.",
+    "static_cover_print_speed_mm_per_sec": "Defines the specific speed at which covers are printed. Set by default to 3/4 of the general printing speed.",
+    "static_cover_print_speed_mm_per_sec_min": "Based on the previous setting.",
+    "static_cover_print_speed_mm_per_sec_max": "Based on the previous setting.",
+    "static_first_layer_print_speed_mm_per_sec": "Defines the specific speed at which the very first layer is printed. Set by default to 1/3 of the general printing speed.",
+    "static_first_layer_print_speed_mm_per_sec_min": "Based on the previous setting.",
+    "static_first_layer_print_speed_mm_per_sec_max": "Based on the previous setting.",
+    "static_travel_speed_mm_per_sec": "Defines the general travel speed at which the extruder moves when not printing. Set by default to 3 times the general printing speed.",
+    "static_travel_speed_mm_per_sec_min": "Based on the previous setting.",
+    "static_travel_speed_mm_per_sec_max": "Based on the previous setting.",
+    "static_acceleration_settings": "Optional and advanced settings. Help modify the way the printer handles acceleration.",
+    "static_enable_acceleration": "Enable / disable acceleration handling. Highly dependant on the firmware and still experimental. May not work as expected on every printer.",
+    "static_x_max_speed": "Maximum speed reachable by the machine on the x-axis. Usually the same as the y-axis.",
+    "static_y_max_speed": "Maximum speed reachable by the machine on the y-axis. Usually the same as the x-axis.",
+    "static_z_max_speed": "Maximum speed reachable by the machine on the z-axis. Usually way lower that the two previous maximums.",
+    "static_e_max_speed": "Maximum speed reachable by the machine idk i'm making things up rn.",
+    "static_x_max_acc": "Max acceleration on the x-axis (mm/s^2). By default, set to 'default acceleration's' value.",
+    "static_y_max_acc": "Max acceleration on the y-axis (mm/s^2). By default, set to 'default acceleration's' value.",
+    "static_z_max_acc": "Max acceleration on the z-axis (mm/s^2). By default, set to 'default acceleration's' value.",
+    "static_e_max_acc": "Max acceleration on the e-axis (mm/s^2). By default, set to 'default acceleration's' value.",
+    "static_default_acc": "Defines the general acceleration (mm/s^2).",
+    "static_e_prime_max_acc": "Defines the acceleration aimed when retracting the filament. Set by default to previous setting's value.",
+    "static_perimeter_acc": "Defines the acceleration aimed when printing the perimeter. Set by default to previous setting's value.",
+    "static_infill_acc": "Defines the acceleration aimed when printing the infill. Set by default to previous setting's value.",
+    "static_default_jerk": "Jerk handles sharp 90-degree edges, by setting the maximum instantaneous change in velocity possible. This setting defines the default value of the printer's jerk.",
+    "static_infill_jerk": "Defines jerk value for infills.",
+    "static_misc_default_settings": "Miscellaneous settings.",
+    "static_add_brim": "A brim is an extension to your print's first layer, deposited before the latter. It serves as a massive adhesion improvement and helps at avoiding warping. This setting toggles the addition of a brim.",
+    "static_brim_distance_to_print_mm": "Defines the distance (mm) from your print's base at which the brim is started.",
+    "static_brim_num_contours": "Defines the number of contours to your brim. More contours means a larger and more adhesive brim, but eventually makes it harder to take off the bed.",
+    "static_enable_z_lift": "Z lift allows the nozzle to lift itself up the z-axis before traveling, thus preventing it from inadvertendly smashing onto your print or dragging it away.",
+    "static_z_lift_mm": "Defines the height of the Z lift.",
+    "static_enable_travel_straight": "Travel straight lets the extruder get from a point to another without staying inside the print's parameter as it would normally do. This lessens travel distances but may cause defects and artifacts if the filament is not properly retracted.",
+}
